cdk/pyproject.toml
[tool.poetry]
name = "cdk"
version = "0.1.0"
description = ""
authors = ["katsut"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "^2.82.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
build = '''
poetry install
poetry build -f wheel
poetry export -f requirements.txt --output dist/requirements.txt --without-hashes
poetry run pip install -r dist/requirements.txt --upgrade --only-binary :all: --platform linux-x86_64 --target dist/package/python dist/*.whl
cd dist/package
zip -r -q /dist/layer.zip . -x '*.pyc'
'''